# gin-blog-server/docs/swagger.yaml
definitions:
  handler.AddOrEditCategoryReq:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  handler.AddOrEditLinkReq:
    properties:
      address:
        type: string
      avatar:
        type: string
      id:
        type: integer
      intro:
        type: string
      name:
        type: string
    required:
    - address
    - name
    type: object
...
  handler.LoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  handler.PageResult-model_FriendLink:
    properties:
      page_data:
        description: 分页数据
        items:
          $ref: '#/definitions/model.FriendLink'
        type: array
      page_num:
        description: 每页条数
        type: integer
      page_size:
        description: 上次页数
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
  handler.PageResult-model_Message:
    properties:
      page_data:
        description: 分页数据
        items:
          $ref: '#/definitions/model.Message'
        type: array
      page_num:
        description: 每页条数
        type: integer
      page_size:
        description: 上次页数
        type: integer
      total:
        description: 总条数
        type: integer
    type: object
...
  handler.RegisterReq:
    properties:
      code:
        type: string
      password:
        maxLength: 20
        minLength: 4
        type: string
      username:
        type: string
    required:
    - code
    - password
    - username
    type: object
  handler.Response-any:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-array_model_OperationLog:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
        items:
          $ref: '#/definitions/model.OperationLog'
        type: array
      message:
        description: 响应消息
        type: string
    type: object
...
  handler.Response-handler_PageResult-model_FriendLink:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/handler.PageResult-model_FriendLink'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-handler_PageResult-model_Message:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/handler.PageResult-model_Message'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
  handler.Response-int:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
        type: integer
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-int64:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        description: 响应数据
        type: integer
      message:
        description: 响应消息
        type: string
    type: object
...
  handler.Response-model_Category:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/model.Category'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-model_CommentVO:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/model.CommentVO'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-model_FriendLink:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/model.FriendLink'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
  handler.Response-model_LoginVO:
    properties:
      code:
        description: 业务状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/model.LoginVO'
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
  handler.UpdateReviewReq:
    properties:
      ids:
        items:
          type: integer
        type: array
      is_review:
        maximum: 1
        minimum: 0
        type: integer
    required:
    - is_review
    type: object
  model.Article:
    properties:
      category_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      desc:
        type: string
      id:
        type: integer
      img:
        type: string
      is_delete:
        type: integer
      is_top:
        type: integer
      original_url:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  model.Category:
    properties:
      articles:
        description: 重写外键
        items:
          $ref: '#/definitions/model.Article'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  model.CommentVO:
    properties:
      article_title:
        type: string
      avatar:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_review:
        type: integer
      nickname:
        type: string
      reply_nickname:
        type: string
      type:
        type: integer
    type: object
  model.FriendLink:
    properties:
      address:
        type: string
      avatar:
        type: string
      created_at:
        type: string
      id:
        type: integer
      intro:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  model.LoginVO:
    properties:
      article_like_set:
        description: '点赞 Set: 用于记录用户点赞过的文章, 评论'
        items:
          type: string
        type: array
      avatar:
        type: string
      comment_like_set:
        items:
          type: string
        type: array
      email:
        type: string
      id:
        type: integer
      intro:
        type: string
      ip_address:
        type: string
      ip_source:
        type: string
      last_login_time:
        type: string
      login_type:
        type: integer
      nickname:
        type: string
      token:
        type: string
      user_info_id:
        type: integer
      username:
        type: string
      website:
        type: string
    type: object
  model.Message:
    properties:
      avatar:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      ip_source:
        type: string
      is_review:
        type: integer
      nickname:
        type: string
      speed:
        type: integer
      updated_at:
        type: string
    type: object
  model.OperationLog:
    properties:
      created_at:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      ip_source:
        type: string
      nickname:
        type: string
      opt_desc:
        type: string
      opt_method:
        type: string
      opt_module:
        type: string
      opt_type:
        type: string
      opt_url:
        type: string
      request_method:
        type: string
      request_param:
        type: string
      response_data:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /category:
    delete:
      consumes:
      - application/json
      description: 根据 ID 数组删除留言
      parameters:
      - description: 留言 ID 数组
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-int'
      security:
      - ApiKeyAuth: []
      summary: 删除留言（批量）
      tags:
      - Category
    post:
      consumes:
      - application/json
      description: 添加或修改分类
      parameters:
      - description: 添加或修改分类
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.AddOrEditCategoryReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-model_Category'
      security:
      - ApiKeyAuth: []
      summary: 添加或修改分类
      tags:
      - Category
...
  /code:
    get:
      consumes:
      - application/json
      description: 发送邮箱验证码
      parameters:
      - description: 邮箱
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            type: string
      summary: 发送邮箱验证码
      tags:
      - UserAuth
  /comment:
    delete:
      consumes:
      - application/json
      description: 根据 ID 数组删除评论
      parameters:
      - description: 评论 ID 数组
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-int'
      security:
      - ApiKeyAuth: []
      summary: 删除评论（批量）
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: 根据条件查询评论列表
      parameters:
      - description: 昵称
        in: query
        name: nickname
        type: string
      - description: 审核状态
        in: query
        name: is_review
        type: integer
      - description: 评论类型
        in: query
        name: type
        type: integer
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-model_CommentVO'
      security:
      - ApiKeyAuth: []
      summary: 条件查询评论列表
      tags:
      - Comment
  /comment/review:
    put:
      consumes:
      - application/json
      description: 根据 ID 数组修改审核状态
      parameters:
      - description: 修改审核状态
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateReviewReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-any'
      security:
      - ApiKeyAuth: []
      summary: 修改评论审核（批量）
      tags:
      - Comment
...
  /link:
    delete:
      consumes:
      - application/json
      description: 根据 ID 数组删除友链
      parameters:
      - description: 友链ID数组
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-int64'
      security:
      - ApiKeyAuth: []
      summary: 删除友链（批量）
      tags:
      - Link
    post:
      consumes:
      - application/json
      description: 添加或修改友链
      parameters:
      - description: 添加或修改友链
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.AddOrEditLinkReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-model_FriendLink'
      security:
      - ApiKeyAuth: []
      summary: 添加或修改友链
      tags:
      - Link
  /link/list:
    get:
      consumes:
      - application/json
      description: 根据条件查询获取友链列表
      parameters:
      - description: 当前页数
        in: query
        name: page_size
        type: integer
      - description: 每页条数
        in: query
        name: page_num
        type: integer
      - description: 搜索关键字
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-handler_PageResult-model_FriendLink'
      security:
      - ApiKeyAuth: []
      summary: 获取友链列表
      tags:
      - Link
  /login:
    post:
      consumes:
      - application/json
      description: 登录
      parameters:
      - description: 登录
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.LoginReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-model_LoginVO'
      summary: 登录
      tags:
      - UserAuth
...
  /message/list:
    get:
      consumes:
      - application/json
      description: 根据条件查询留言列表
      parameters:
      - description: 昵称
        in: query
        name: nickname
        type: string
      - description: 审核状态
        in: query
        name: is_review
        type: integer
      - description: 当前页数
        in: query
        name: page_size
        type: integer
      - description: 每页条数
        in: query
        name: page_num
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-handler_PageResult-model_Message'
      security:
      - ApiKeyAuth: []
      summary: 条件查询留言列表
      tags:
      - Message
  /message/review:
    put:
      consumes:
      - application/json
      description: 根据 ID 数组修改审核状态
      parameters:
      - description: 修改审核状态
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateReviewReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-int'
      security:
      - ApiKeyAuth: []
      summary: 修改留言审核（批量）
      tags:
      - Message
  /operation/log:
    delete:
      consumes:
      - application/json
      description: 删除操作日志
      parameters:
      - description: 操作日志ID列表
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-int'
      security:
      - ApiKeyAuth: []
      summary: 删除操作日志
      tags:
      - OperationLog
  /operation/log/list:
    get:
      consumes:
      - application/json
      description: 根据条件查询获取操作日志列表
      parameters:
      - description: 页码
        in: query
        name: page_num
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            $ref: '#/definitions/handler.Response-array_model_OperationLog'
      security:
      - ApiKeyAuth: []
      summary: 获取操作日志列表
      tags:
      - OperationLog
...
  /register:
    post:
      consumes:
      - application/json
      description: 注册
      parameters:
      - description: 注册
        in: body
        name: form
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterReq'
      produces:
      - application/json
      responses:
        "0":
          description: ""
          schema:
            type: string
      summary: 注册
      tags:
      - UserAuth
...
swagger: "2.0"